Do not skip MzMLCV.cvTIC
Some vendors report a different TIC CV param value from the one obtained from integrating from the scan's intensities.
This may be problematic if the TIC CV param value is the result e.g. data processing and should be used for downstream analysis.
Retain the MzMLCV.cvTIC so it can be retrieved from the MzMLMsScan object manually, if necessary.